develop findAll by pageable

+++ src/main/java/ir/maktabsharif101/jpaexample/base/repository/Pageable.java
@@ -4,7 +4,7 @@ public interface Pageable {
 
     int getPageNumber();
 
-    long getPageSize();
+    int getPageSize();
 
     long getOffset();
